+++ lib/ecto/date_time_range.ex
@@ -20,17 +20,27 @@ defmodule Ecto.DateTimeRange do
   iex> ~t(2020-02-02T00:01:00Z..2020-02-02T00:01:01Z)U
   %Ecto.DateTimeRange.UTCDateTime{start_at: ~U[2020-02-02T00:01:00Z], end_at: ~U[2020-02-02T00:01:01Z]}
   ...>
-  iex> ~t(hi there)
-  ** (ArgumentError) Unable to parse DateTime(s) from input
+  iex> ~t(2020-02-02T00:01:00..2020-02-02T00:01:01)N
+  %Ecto.DateTimeRange.NaiveDateTime{start_at: ~N[2020-02-02T00:01:00], end_at: ~N[2020-02-02T00:01:01]}
   ...>
   iex> ~t(00:01:00..00:01:01)T
   %Ecto.DateTimeRange.Time{start_at: ~T[00:01:00], end_at: ~T[00:01:01]}
+  ...>
+  iex> ~t(hi there)
+  ** (ArgumentError) Unable to parse DateTime(s) from input
   ```
   """
   def sigil_t(string, []), do: sigil_t(string, [?U])
 
   def sigil_t(string, [?U]) when is_binary(string) do
     case Ecto.DateTimeRange.UTCDateTime.parse(string) do
+      {:ok, range} -> range
+      {:error, error} -> raise ArgumentError, error
+    end
+  end
+
+  def sigil_t(string, [?N]) when is_binary(string) do
+    case Ecto.DateTimeRange.NaiveDateTime.parse(string) do
       {:ok, range} -> range
       {:error, error} -> raise ArgumentError, error
     end